+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS23X26HBQCMT1773EH7W6JB"
+        "01KS25M280EQRT4YW3GX6CA24M"
     )
     r"""The unique identifier of the data source"""